merge

+++ src/java.desktop/share/classes/javax/swing/text/WrappedPlainView.java
@@ -360,7 +360,8 @@ public class WrappedPlainView extends BoxView implements TabExpander {
         int currentWidth = getWidth();
         if (wordWrap) {
             p = p0 + Utilities.getBreakLocation(segment, metrics,
-                                                tabBase, tabBase + currentWidth,
+                                                (float)tabBase,
+                                                (float)(tabBase + currentWidth),
                                                 this, p0);
         } else {
             p = p0 + Utilities.getTabbedTextOffset(segment, metrics,
@@ -465,7 +466,7 @@ public class WrappedPlainView extends BoxView implements TabExpander {
     public float nextTabStop(float x, int tabOffset) {
         if (tabSize == 0)
             return x;
-        float ntabs = (x - tabBase) / tabSize;
+        int ntabs = (int) ((x - tabBase) / tabSize);
         return tabBase + ((ntabs + 1) * tabSize);
     }
